WIP Target DTL in a restricted (non-stateful sense)
This approach was attempted because it represented a minimal way to get
PyOP3 lowering to DTL. However, local kernels are supposed to be more
generic than this and so either the monad approach (too invasive) or
something non-DTL is probably required. I would still like to get
assembly kernels using just DTL though. It would be CEED-like.