WIP
So I think that I finally have the right way of expressing the different
ways of indexing and slicing a tensor. The only issue I am now facing is
the fact that I have done away with 'domains' where a domain is just a
start, stop tuple. I need some way to store this information s.t. we
don't get any duplication.