About to try something quite different
The problem:
Collecting shared inames is a big headache. Indices may be different on
the LHS and RHS of the assignment so finding what the index expression
for each side and determining 'within_inames' is really hard.

The solution:
The LHS and RHS of the assignment must have the same *indexed* shape,
that is, the 'within_loops' do not matter but all the others do. This
means that all of the loops 'below' the 'within_loops' are shared.

My idea is to, for each assignment, create a stack of loops/inames
based upon the indexed shapes *before anything else* and then pass this
to the expression constructors of the LHS and RHS.