Some things work but others don't
In particular I have found that for this to work I need to uniquify
things some more as dim labels can crop up all over the place. In
particular I am facing a challenge for mixed temporaries where the same
dim label is used for different bits of it. I will switch to storing the
subdim_id most of the time and only use dim labels for iname
determination.