New tree approach
The main idea is that of a "labelled" tree. This seems to encapsulate
best what my multi-axis stuff is all about. It should make creating a
multi-axis more straightforward but still have all the transformation
benefits of the new tree classes.

I do, for instance, think that the syntax is now much cleaner. Slowly
the algorithms are becoming less mysterious...