WIP Begin refactoring axis setup
Currently all the same tests are passing.

I am opting for a much simpler approach to determining layouts. In
particular I am restricting permutations to only be for single axes and
not be nestable or work with ragged. I cannot think of a use case for
needing anything else and I also don't think it would be terribly hard
to add back in.

I am also not pushing data layouts to the bottom of the stack. It didn't
seem worth it and made things even more confusing.