Can tabulate (basic) layouts that rely on an external loop index
I am not totally confident in my approach, but the layouts do appear
right.

A particular question of mine is what we should do for axes higher up
the chain. The internal step sizes are now non-const. Perhaps this is
fine we should just avoid determining the steps when axis==axes.root.