WIP, about to refactor index_exprs
Most tests currently pass but I'm about to blow that all away.

I think that index_exprs should be split into "outer_loops" and
index_exprs. This is because the current policy for identifying outer
loops is to use {None: {axis_label: loop_index_var, ...}}. This
is insufficient because if we have multiple loop indices they will
clash.

To resolve this I will try collecting loop index variables in a
frozenset "outer_loops" attribute instead of storing as None.