WIP, add DummyKernelArgument and NA intent
Trying to generate code without needing the data structures to exist.
The generated code looks right for the moment. The final piece for the
single-cell wrapper (I think) is to enable one to pass loop indices in
as arguments. I think all that is required for that to work is to do a
bit extra when we unwrap loop contexts. Loop indices should be arrays of
integers rather than single ints because we can happily pass around
multi-indices.