Some diabolical hackery
The key idea is that for us to be able to optimise access patterns we
need to more eagerly generate the index expressions (`subst_layouts`)
that we use in codegen. Therefore we want to deal with 'concrete' arrays
that have established layout functions.

We therefore have `_LayoutDat` and `_ConcretizedDat` that represent
these. I eagerly construct `_LayoutDat` when I index because arrays in
expressions aren't really indexable any more. This grates though and I
should probably only do this at code generation time.

Now we are dealing with 'concrete' arrays it is not tractable to
optimise layouts.